+++ packages/vue/src/types.ts
@@ -2,6 +2,7 @@
  * Type to make properties optional and preserve their type
  */
 export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>
+export type Assign<T, U> = Omit<T, keyof U> & U
 
 export type EmitFn<T> = <K extends keyof T>(
   event: K,