+++ packages/vue/CHANGELOG.md
@@ -6,6 +6,8 @@ description: All notable changes will be documented in this file.
 
 ## [Unreleased]
 
+## [1.0.0-6] - 2024-05-22
+
 ### Added
 
 - Exported `Assign` and `Optional` types.